Moved Task interface to interfaces folder.
